Revert "AB#25833"
