+++ package.json
@@ -67,7 +67,6 @@
     "node-cache": "^5.1.2",
     "node-cron": "^3.0.0",
     "node-fetch": "^2.6.1",
-    "nodemailer": "^6.7.2",
     "pascal-case": "^3.1.2",
     "passport": "^0.4.1",
     "passport-azure-ad": "^4.2.1",
