--- src/const/placeholders.ts
@@ -1,6 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/naming-convention
-export enum Placeholders {
-  TODAY = '{today}',
-  DATASET = '{dataset}',
-  NOW = '{now}',
-}
