+++ src/routes/download/index.ts
@@ -18,10 +18,13 @@ import {
   templateBuilder,
   getColumns,
   getRows,
-  extractGridData,
+  getColumnsFromMeta,
+  getRowsFromMeta,
 } from '../../utils/files';
 import sanitize from 'sanitize-filename';
 import mongoose from 'mongoose';
+import { buildQuery, buildMetaQuery } from '../../utils/query/queryBuilder';
+import fetch from 'node-fetch';
 
 /**
  * Exports files in csv or xlsx format, excepted if specified otherwised
@@ -169,12 +172,10 @@ router.get('/resource/records/:id', async (req, res) => {
  * The parameters are :
  * params = {
  *    ids?: string[],                     // If exportOptions.records === 'selected', list of ids of the records
+ *    resId: number,
  *    fields?: any[],                     // If exportOptions.fields === 'displayed', list of the names of the fields we want to export
  *    filter?: any                        // If any set, list of the filters we want to apply
- *    format: 'csv' | 'xlsx'              // Export on csv or excel format
- *    query: any                          // Query parameters to build it
- *    sortField?: string
- *    sortOrder?: 'asc' | 'desc'
+ *    format: 'csv' | 'xlsx'           // Export on csv or excel format
  * }
  */
 router.post('/records', async (req, res) => {
@@ -184,9 +185,68 @@ router.post('/records', async (req, res) => {
     return res.status(400).send('Missing parameters');
   }
 
-  const { columns, rows } = await extractGridData(
-    params,
-    req.headers.authorization
+  const query = buildQuery(params.query);
+  const metaQuery = buildMetaQuery(params.query);
+
+  let records: any[] = [];
+  let meta: any;
+
+  const gqlQuery = fetch('http://localhost:3000/graphql', {
+    method: 'POST',
+    body: JSON.stringify({
+      query: query,
+      variables: {
+        first: 5000,
+        sortField: params.sortField,
+        sortOrder: params.sortOrder,
+        filter: params.filter,
+        display: true,
+      },
+    }),
+    headers: {
+      Authorization: req.headers.authorization,
+      'Content-Type': 'application/json',
+    },
+  })
+    .then((x) => x.json())
+    .then((y) => {
+      for (const field in y.data) {
+        if (Object.prototype.hasOwnProperty.call(y.data, field)) {
+          records = y.data[field].edges.map((x) => x.node);
+        }
+      }
+    });
+
+  const gqlMetaQuery = fetch('http://localhost:3000/graphql', {
+    method: 'POST',
+    body: JSON.stringify({
+      query: metaQuery,
+    }),
+    headers: {
+      Authorization: req.headers.authorization,
+      'Content-Type': 'application/json',
+    },
+  })
+    .then((x) => x.json())
+    .then((y) => {
+      for (const field in y.data) {
+        if (Object.prototype.hasOwnProperty.call(y.data, field)) {
+          meta = y.data[field];
+        }
+      }
+    });
+
+  await Promise.all([gqlQuery, gqlMetaQuery]);
+
+  const rawColumns = getColumnsFromMeta(meta);
+  const columns = rawColumns.filter((x) =>
+    params.fields.find((y) => y.name === x.name)
+  );
+  const rows = await getRowsFromMeta(columns, records);
+
+  // Edits the column to match with the fields
+  columns.forEach(
+    (x) => (x.title = params.fields.find((y) => y.name === x.name).title)
   );
 
   // Returns the file
