+++ src/server/apollo/context.ts
@@ -4,7 +4,6 @@ import { AppAbility } from 'security/defineAbilityFor';
 export interface Context {
   user: UserWithAbility;
   dataSources?: any;
-  token?: string;
 }
 
 interface UserWithAbility extends User {
@@ -14,13 +13,11 @@ interface UserWithAbility extends User {
 export default ({ req, connection }): Context => {
   if (connection) {
     return {
-      token: req.headers.authorization,
       user: connection.context.user,
     } as Context;
   }
   if (req) {
     return {
-      token: req.headers.authorization,
       // not a clean fix but that works for now
       user: (req as any).user,
     } as Context;
