+++ src/utils/files/index.ts
@@ -8,5 +8,3 @@ export * from './getRows';
 export * from './loadRow';
 export * from './getColumnsFromMeta';
 export * from './getRowsFromMeta';
-export * from './extractGridData';
-export * from './emailBuilder';
