+++ src/utils/files/xlsBuilder.ts
@@ -4,12 +4,13 @@ import get from 'lodash/get';
 /**
  * Builds an XLSX file.
  *
+ * @param res Request response
  * @param fileName Name of the file
  * @param columns Array of objects with a name property that will match the data, and optionally a label that will be the column title on the exported file
  * @param data Array of objects, that will be transformed into the rows of the csv. Each object should have [key, value] as [column's name, corresponding value].
  * @returns response with file attached.
  */
-export default (fileName: string, columns: any[], data) => {
+export default async (res, fileName: string, columns: any[], data) => {
   const workbook = new Workbook();
   const worksheet = workbook.addWorksheet(fileName);
 
@@ -40,6 +41,17 @@ export default (fileName: string, columns: any[], data) => {
     worksheet.addRow(temp);
   }
 
+  // Set response parameters
+  res.setHeader(
+    'Content-Type',
+    'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+  );
+  res.setHeader(
+    'Content-Disposition',
+    'attachment; filename=' + `${fileName}.xlsx`
+  );
+
   // Write to a new buffer
-  return workbook.xlsx.writeBuffer();
+  const buffer = await workbook.xlsx.writeBuffer();
+  return res.send(buffer);
 };
